+++ sponsoredTransactionsAuction/frontend/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0
 
-- Initial auction sponsored transaction front end
+-   Initial auction sponsored transaction front end